azure-functions/requirements.txt
azure-storage-blob==12.26.0
azure-functions==1.23.0